:art: factorize some code
