+++ src/types/simple/simple.selectors.js
@@ -1,29 +1,16 @@
+import { getState } from '../../selectors'
 import { initState } from './simple.middleware'
 
-// TODO: make it a util (and factorize code)
-const getFromPath = (data, path) => path.split('.').reduce(
-  (curr, sub) => curr && curr[sub],
-  data,
-)
-
 export default (options) => {
-  // TODO: make it a util (and factorize code)
-  const getState = (rootState) => {
-    const { path, name } = options
-    let subState = rootState
-
-    if (path !== undefined && path.length > 0) subState = getFromPath(rootState, path)
-
-    return subState[name]
+  const getStateWithOptions = getState(options)
+  const get = () => rootState => getStateWithOptions(rootState)
+  const isInitialized = (rootState) => {
+    if (options.defaultData !== undefined) return getStateWithOptions(rootState) !== options.defaultData
+    return get()(rootState) !== initState
   }
-
-  const get = () => rootState => getState(rootState)
 
   return {
     get,
-    isInitialized: (rootState) => {
-      if (options.defaultData !== undefined) return getState(rootState) !== options.defaultData
-      return get()(rootState) !== initState
-    },
+    isInitialized,
   }
 }
